Rename OppdaterAktivitetEtikett til tsx
