+++ src/moduler/aktivitet/visning/etikett-oppdatering/OppdaterAktivitetEtikett.tsx
